Fix uninitialized Vec2 members, missing SFX, size_t underflow
Bug fixes from deep code analysis:
- TransformComponent: default-init position + prevPosition to {0,0}
  (fixes garbage in getInterpolatedRect on default-constructed entities)
- AIComponent: default-init patrolStart, patrolEnd, bossTelegraphDir,
  targetPosition to {0,0}
- Add SFX to right-click-back in LoreState, PauseState, TutorialState,
  UpgradeState (4 states were silent on right-click)
- Add SFX to TutorialState mouse click (left/right page navigation)
- Fix size_t underflow in OptionsState: m_options.size()-3 could wrap
  to huge value if vector had <3 elements (cast to int)

+++ src/States/OptionsState.cpp
@@ -162,7 +162,7 @@ void OptionsState::handleEvent(const SDL_Event& event) {
                     m_options[OPT_HUDSCALE].value   = 100; // HUD scale 100%
                     m_options[OPT_LANGUAGE].value   = 0;   // English
                     m_options[OPT_CRT].value        = 0;   // CRT off
-                    for (size_t i = 0; i < m_options.size() - 3; i++) applyOption(static_cast<int>(i));
+                    for (int i = 0; i < static_cast<int>(m_options.size()) - 3; i++) applyOption(i);
                     game->saveSettings();
                     AudioManager::instance().play(SFX::MenuConfirm);
                     // Re-enter to refresh labels after language reset
@@ -385,7 +385,7 @@ void OptionsState::update(float dt) {
             m_options[OPT_HUDSCALE].value   = 100;
             m_options[OPT_LANGUAGE].value    = 0;
             m_options[OPT_CRT].value         = 0;
-            for (size_t i = 0; i < m_options.size() - 3; i++) applyOption(static_cast<int>(i));
+            for (int i = 0; i < static_cast<int>(m_options.size()) - 3; i++) applyOption(i);
             game->saveSettings();
             AudioManager::instance().play(SFX::MenuConfirm);
             { int sel = m_selected; enter(); m_selected = sel; }
